requirements.txt
streamlit
python-dotenv
pymysql
alembic
cryptography
pytest
 
#requests
#pandas
#yfinance
#numpy
#jupyterlab
#matplotlib
#scipy